# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: helmstudio platform API
  version: "1.0.0"
  description: >
    The local, single-machine API a studio's runtime SDK speaks, and the one
    the launcher speaks to the daemon. Served by the daemon at
    127.0.0.1:8700 and, for one studio with no daemon, by `helm dev` and the
    embedded provider (05 §2, §5a). The two are substitutable: one
    conformance suite runs against both (Q25). The in-process embedded
    provider has no token, and still enforces the studio's capabilities,
    read from its manifest, with the same 403 `capability_required`; the
    suite's capability cases run against both providers (M4 first review,
    #7).
servers:
  - url: http://127.0.0.1:8700/api/v1
...
security: []
paths:
...
  /assets:adopt:
    post:
      tags: [studio-api]
      operationId: assetsAdopt
      x-helm-group: assets
      x-helm-method: adopt
      x-helm-capability: assets
      summary: Adopt a file from the stage or data directory by hardlink — no copy, no HTTP body.
      description: >
        R34, 07 §4, Q10 as amended by the M4 first review (#4), Q11.

        **Where the file may be.** `path` must resolve, with no symlink in it,
        inside the caller's stage directory (`HELM_STAGE_DIR`, `paths.stage`)
        or its data directory (`{data}`, `paths.data`). Anywhere else is 422
        `outside_roots`. It must be a regular file (422 `not_a_file`).

        **What happens to it.** The file is hardlinked into the blob store and
        made read-only (0444).
        - From stage, the stage entry is then unlinked.
        - From data, the studio's file stays where it is. It is the blob's
          inode, so it is read-only too.

        **When the bytes already exist.** Nothing is stored and the existing
        asset returns with 200 (02 §7). A stage entry is unlinked; a data file
        is left alone.

        **Adoption never copies.** A blob hardlink that fails across volumes is
        422 `cross_device`, and the file is left in place. A library hardlink
        that fails across volumes still adopts, with `library_path` null and a
        warning.

        An asset stays out of reclaim until an item or item input first
        references it (M4 first review, #3).
      security: [{ studioToken: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/AdoptRequest" } } }
      responses:
        "201":
          description: A new asset.
          content: { application/json: { schema: { $ref: "#/components/schemas/Asset" } } }
        "200":
          description: These bytes were already stored; the existing asset.
          content: { application/json: { schema: { $ref: "#/components/schemas/Asset" } } }
        "400": { $ref: "#/components/responses/Invalid" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/Unprocessable" }
...
  /assets:reclaim:
    get:
      tags: [launcher]
      operationId: assetsReclaimPreview
      x-helm-group: assets
      x-helm-method: reclaimPreview
      summary: Exactly which assets reclaim would delete, and the bytes freed.
      description: >
        Q12, 02 §8, M4 first review #1 and #3. An asset is reclaimable when
        all of these hold:

        - It is unpinned.
        - It was referenced by an item or item input at least once.
        - No live (not soft-deleted) item has it as its asset or as an input.
        - No timeline clip names it.

        `total_bytes` counts only assets whose stored file has no link outside
        the asset store and library: an asset adopted from a studio's `{data}`
        is still listed, and frees nothing while the studio keeps its file
        (second review #7). Each item in the preview lists the soft-deleted
        items that name the asset. Reclaim hard-deletes those, because until then a soft delete is
        undoable. Assets never referenced are counted in `unreferenced`, and
        never deleted. `confirm` is a digest of the exact set, including
        those items.
      responses:
        "200":
          description: The preview.
          content: { application/json: { schema: { $ref: "#/components/schemas/AssetReclaimPreview" } } }
    post:
      tags: [launcher]
      operationId: assetsReclaim
      x-helm-group: assets
      x-helm-method: reclaim
      summary: Delete exactly the previewed set.
      description: >
        Q12, 02 §8. When `confirm` no longer matches the current preview,
        nothing is deleted: 409 `preview_changed`, with the current preview in
        `details.preview`. Otherwise, in one transaction:

        1. Hard-delete every soft-deleted item listed. Its inputs, tags, inbox
           entries and search entry go with it.
        2. Delete the asset rows.

        After that commits, reclaim removes each asset's blob and derived
        directory. It removes the library file only when that path still holds
        the blob's inode; a file the user replaced there is left in place and
        reported in `kept_library_paths`.

        A soft-deleted item never blocks reclaim with a foreign-key error.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ReclaimConfirm" } } }
      responses:
        "200":
          description: What was deleted.
          content: { application/json: { schema: { $ref: "#/components/schemas/AssetReclaimResult" } } }
        "400": { $ref: "#/components/responses/Invalid" }
        "409": { $ref: "#/components/responses/Conflict" }

  # ======================================================================
  # Gallery (R38, 07 §5)
  # ======================================================================
...
  /studios:
    get:
      tags: [launcher]
      operationId: studiosList
      x-helm-group: studios
      x-helm-method: list
      summary: The library — every studio this machine knows about, from every source, ordered by id.
      description: >
        R3b, M7 Q6 and Q18. The library is the set of studios this machine
        knows about, of which installed ones are a subset. One entry per id,
        decided by the highest source that has a file for it — local, the
        repository, then the registry pointer's inline manifest — and decided
        by **existence, not validity**: an invalid winner is listed invalid,
        with its source, its file and its errors, and the lower sources are not
        read in its place. Every entry carries `source`, and `overrides` when
        it shadows another.

        Includes studios still running from a previous daemon whose manifest
        is gone or invalid (`manifest_loaded: false`; M2 review round 1).

        Nothing here touches the network: startup reads the bundled registry,
        the local directory, installed checkouts and the manifest cache, and a
        pointer with nothing cached is listed with `manifest_state: not_fetched`
        (Q7).
      parameters:
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Cursor" }
      responses:
        "200":
          description: A page of studios.
          content: { application/json: { schema: { $ref: "#/components/schemas/StudioPage" } } }
        "400": { $ref: "#/components/responses/Invalid" }
  /studios/{id}:
    get:
      tags: [launcher]
      operationId: studiosGet
      x-helm-group: studios
      x-helm-method: get
      summary: One studio.
      parameters: [{ $ref: "#/components/parameters/StudioId" }]
      responses:
        "200":
          description: The studio.
          content: { application/json: { schema: { $ref: "#/components/schemas/Studio" } } }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /studios/{id}/processes:
    get:
      tags: [launcher]
      operationId: studiosProcesses
      x-helm-group: studios
      x-helm-method: processes
      summary: The studio's current or most recent process group.
      parameters: [{ $ref: "#/components/parameters/StudioId" }]
      responses:
        "200":
          description: The group.
          content: { application/json: { schema: { $ref: "#/components/schemas/ProcessGroup" } } }
        "404": { $ref: "#/components/responses/NotFound" }
  /studios/{id}/logs:
    get:
      tags: [launcher]
      operationId: studiosLogFiles
      x-helm-group: studios
      x-helm-method: logFiles
      summary: The studio's retained run logs, newest first.
      description: R30. The last five group runs per studio.
      parameters:
        - { $ref: "#/components/parameters/StudioId" }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Cursor" }
      responses:
        "200":
          description: A page of log files.
          content: { application/json: { schema: { $ref: "#/components/schemas/LogFilePage" } } }
        "400": { $ref: "#/components/responses/Invalid" }
        "404": { $ref: "#/components/responses/NotFound" }
  /studios/{id}/processes/{name}/logs:
    get:
      tags: [launcher]
      operationId: studiosProcessLog
      x-helm-group: studios
      x-helm-method: processLog
      summary: One process's output as server-sent events.
      description: >
        docs/decisions.md M2. Events: `line` (id: its sequence), `gap` (lines
        this viewer lost for being slow, or bytes the daemon skipped), `end`
        when the process is not live and its tail has been sent. Resume with
        `Last-Event-ID`.
      parameters:
        - { $ref: "#/components/parameters/StudioId" }
        - name: name
          in: path
          required: true
          schema: { type: string }
        - name: Last-Event-ID
          in: header
          schema: { type: string }
      x-helm-events:
        line: { $ref: "#/components/schemas/ProcessLogLine" }
        gap: { $ref: "#/components/schemas/ProcessLogLine" }
        end: { $ref: "#/components/schemas/ProcessLogEnd" }
      responses:
        "200":
          description: An event stream.
          content: { text/event-stream: { schema: { type: string } } }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /studios/{id}/approval:
    get:
      tags: [launcher]
      operationId: studiosApproval
      x-helm-group: studios
      x-helm-method: approval
      summary: What would run, and the digest that authorises it.
      description: >
        R62, R63, Q10-Q13. Reads the repository and resolves `ref` to a commit
        first, so the preview shows the source that will actually be used, then
        returns every command grouped by when it runs, the submodules, the
        weights with sizes, the capabilities as sentences, the declared hosts
        and the derived level.

        **Only checks that execute nothing run.** Theme conformance and the
        smoke test are listed as not run, with the reason, never as passes.

        `digest` covers everything that executes and where it comes from: the
        id, source and resolved commit, the submodules, every command with its
        `cwd`, `shell` and `env`, the capabilities, the network hosts,
        `python.version` and the declared weights. It does **not** cover sizes,
        free disk or the checkpoint selection (Q21).

        Reading a repository needs the network; a repository that cannot be
        reached is 503 with what is known.
      parameters: [{ $ref: "#/components/parameters/StudioId" }]
      responses:
        "200":
          description: The preview.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApprovalPreview" } } }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/Unprocessable" }
        "503": { $ref: "#/components/responses/Unavailable" }
...
components:
  securitySchemes:
    studioToken:
      type: http
      scheme: bearer
      description: >
        HELM_TOKEN, minted per launch for one studio and its declared
        capabilities, revoked when its group stops (07 §3, Q6). Format
        `hs_live_` plus 52 base32 characters.
...
  responses:
    Invalid:
      description: >
        The request is malformed. Codes: `bad_request`, `bad_filter`,
        `bad_cursor`, `invalid_document`.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unauthenticated:
      description: No studio token, or a revoked or unknown one. Code `unauthenticated`.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Forbidden:
      description: >
        The token lacks a capability (`capability_required`, with
        `details.capability`), the job is not a task the studio may change
        (`not_a_task`), or the request came from another origin (`bad_origin`).
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Absent, deleted, or not the caller's to see. Code `not_found`.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    Unprocessable:
      description: >
        Well-formed but refused. Codes: `outside_roots`, `cross_device`,
        `not_a_file`, `unknown_session`, `self_handoff`, `not_launchable`,
        `blocked`, `not_linkable`.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
...
    Asset:
      type: object
      required: [id, sha256, kind, mime, bytes, width, height, duration_s, fps, state, pinned, origin_studio, url, thumb, library_path, created_at]
      properties:
        id: { $ref: "#/components/schemas/ULID" }
        sha256: { type: string, pattern: "^[0-9a-f]{64}$" }
        kind: { $ref: "#/components/schemas/AssetKind" }
        mime: { type: string }
        bytes: { type: integer }
        width: { type: [integer, "null"] }
        height: { type: [integer, "null"] }
        duration_s: { type: [number, "null"] }
        fps: { type: [number, "null"] }
        state: { type: string, enum: [ready, missing, corrupt] }
        pinned: { type: boolean }
        origin_studio:
          type: [string, "null"]
          description: >
            The studio that first stored these bytes. Null for any caller that
            is not that studio, so a dedup hit or a shared read never reveals
            another studio (M4 first review, #9).
        url: { type: string, description: "This API's path to the bytes, e.g. /api/v1/assets/{id}. Needs the Bearer header; a page prefixes /helm to go through its proxy (M6 Q10)." }
        thumb: { type: string, description: "This API's path to the thumbnail, e.g. /api/v1/assets/{id}/thumb?w=320. Needs the Bearer header; a page prefixes /helm to go through its proxy (M6 Q10)." }
        library_path:
          type: [string, "null"]
          description: >
            The readable hardlink, relative to the library root. Null when it
            could not be made (Q11), and null for any caller that is not the
            origin studio, since it names that studio, a month and a filename
            (M4 first review, #9).
        created_at: { type: string, format: date-time }
        warnings:
          type: array
          items: { type: string }
          description: Present on an adopt or upload response when something short of failure happened, such as no library link.
    AdoptRequest:
      type: object
      required: [path, kind]
      additionalProperties: false
      properties:
        path: { type: string, description: "Absolute, inside the caller's stage directory (unlinked after adopting) or its data directory (left in place, read-only)." }
        kind: { $ref: "#/components/schemas/AssetKind" }
        pinned: { type: boolean, default: false }
        width: { type: integer, minimum: 1 }
        height: { type: integer, minimum: 1 }
        duration_s: { type: number, exclusiveMinimum: 0 }
        fps: { type: number, exclusiveMinimum: 0 }
...
    AssetReclaimPreview:
      type: object
      required: [items, total_bytes, unreferenced, confirm]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/AssetReclaimItem" } }
        total_bytes: { type: integer }
        unreferenced:
          type: object
          required: [count, bytes]
          description: Unpinned assets no item or input has ever referenced. Reported, never reclaimed.
          properties:
            count: { type: integer }
            bytes: { type: integer }
        confirm: { type: string }
    AssetReclaimResult:
      type: object
      required: [items, total_bytes, kept_library_paths]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/AssetReclaimItem" } }
        total_bytes: { type: integer }
        kept_library_paths:
          type: array
          items: { type: string }
          description: Library paths left in place because they no longer held the blob's inode.
    ReclaimConfirm:
      type: object
      required: [confirm]
      additionalProperties: false
      properties:
        confirm: { type: string }

    # ---------------------------------------------------------------- gallery
...